refactored routes from app.js file so that they are each in their own grouped files
