+++ routes/campgrounds.js
@@ -1,0 +1,62 @@
+const express = require("express");
+const router = express.Router();
+
+const Campground = require("../models/campground");
+
+// CAMPGROUNDS PAGE
+// ======================================================
+router.get("/", function(req, res) {
+    // Get all campgrounds from the DB
+    Campground.find({}, function(err, allCampgrounds) {
+        if (err) {
+            console.log(err)
+        } else {
+            res.render("campgrounds/index", {campgrounds: allCampgrounds});
+        }
+    })
+    
+});
+
+// CAMGROUNDS NEW PAGE
+// ======================================================
+router.get("/new", function(req, res) {
+    res.render("campgrounds/new")
+});
+
+
+// CAMGROUNDS CREATE ACTION
+// ======================================================
+router.post("/", function(req, res) {
+    // get data from the form and add to campgrounds array
+    const name = req.body.name;
+    const image = req.body.image;
+    const desc = req.body.description;
+    const newCampground = {name: name, image: image, description: desc};
+    
+    // Create a new campground and save to database
+    Campground.create(newCampground, function(err, newlyCreated) {
+        if (err) {
+            console.log(err);
+        } else {
+            res.redirect("/campgrounds");
+        }
+    })
+});
+
+// CAMGROUNDS SHOW PAGE
+// ======================================================
+router.get("/:id", function(req, res) {
+    // Find the campground with the provided ID
+    Campground.findById(req.params.id).populate("comments").exec(function(err, foundCampground) {
+        if(err) {
+            console.log(err);
+        } else {
+            console.log(foundCampground);
+            // render show template for that campground
+            res.render("campgrounds/show", {campground: foundCampground});
+        }
+    });
+});
+
+
+module.exports = router;
